% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTargetTable.R
\name{createTargetTable}
\alias{createTargetTable}
\title{Build the study target table}
\usage{
createTargetTable(
  connectionDetails,
  cohortDatabaseSchema,
  cohortTable,
  targetTable,
  targetMetastasis,
  targetBones,
  bpaZa,
  bpaDenosumab,
  adt,
  arpi,
  chemo,
  packageName = "PioneerBPA"
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object.}

\item{cohortDatabaseSchema}{Schema holding the cohort and target tables
(needs read / write / delete).}

\item{cohortTable}{Base name of the cohort table populated by
\code{\link{generateStudyCohorts}}.}

\item{targetTable}{Name of the target table to build.}

\item{targetMetastasis}{\code{cohort_definition_id} of the metastatic
prostate cancer cohort.}

\item{targetBones}{\code{cohort_definition_id} of the bone metastatic site cohort.}

\item{bpaZa}{\code{cohort_definition_id} of the zoledronic acid (ZA) cohort.}

\item{bpaDenosumab}{\code{cohort_definition_id} of the denosumab cohort.}

\item{adt}{\code{cohort_definition_id} of the ADT cohort.}

\item{arpi}{\code{cohort_definition_id} of the ARPI cohort.}

\item{chemo}{\code{cohort_definition_id} of the chemotherapy cohort.}

\item{packageName}{Name of the package holding the SQL resources.}
}
\description{
Assembles the study \code{targetTable} from the generated cohorts by running
three combination SQL scripts in sequence: metastasis + bone,
metastasis + bone + BPA, and metastasis + bone + BPA + systemic treatment.
}
